requirements.txt
Flask
Flask-WTF
Flask-Login
Flask-SQLAlchemy
Flask-Migrate
